[Article] update tests
